Rewrote seedfunction and schema to account only for prop_id and url
